06_MCP/requirements.txt
# --- Servidor MCP ---
mcp[server]
fastapi
uvicorn
httpx
requests

# --- Inteligencia Artificial (Groq & LangGraph) ---
groq
langgraph
langchain-groq
python-dotenv
aiohttp

# --- Análisis y Visualización de Datos ---
pandas
grandalf

# --- Interfaces de Usuario ---
streamlit
gradio

